refactor: extract Description into its own file
`Description` had grown to ~55 lines in `Aliases.js` — large enough to
dominate the file, which is intended for small one-class shims. Move it
to `components/Description.js`, default-exported and registered in
`ComponentTypes.js` alongside `BlockQuote`.

`Aliases.js` drops the now-unused
`returnPassThroughListItemChildStateVariableDefinitions` import; the
other shared imports (`BlockComponent`, `InlineComponent`, etc.) are
still in use.

Pure code movement; no behavior change. Targeted tests pass:
`accessibility.test.ts` (8), `schema-enforcement.test.ts` (6).
