+++ packages/doenetml-worker-javascript/src/components/Aliases.js
@@ -9,7 +9,6 @@ import { Div } from "./Divisions";
 import Setup from "./Setup";
 import P from "./P";
 import Case from "./Case";
-import { returnPassThroughListItemChildStateVariableDefinitions } from "../utils/listItemChild";
 
 export class Title extends TextOrInline {
     static componentType = "title";
@@ -27,64 +26,6 @@ export class RightHandSide extends MathComponent {
             "A right-hand-side expression of an equation in an ODE system.",
     };
     static rendererType = "math";
-}
-
-export class Description extends BlockComponent {
-    constructor(args) {
-        super(args);
-
-        Object.assign(this.actions, {
-            recordVisibilityChange: this.recordVisibilityChange.bind(this),
-        });
-    }
-    static componentType = "description";
-
-    static componentDocs = {
-        summary:
-            "Extra information about an enclosing component, shown to all users in a popup or disclosure.",
-    };
-    static rendererType = "containerBlock";
-    static renderChildren = true;
-
-    static canDisplayChildErrors = true;
-
-    static includeBlankStringChildren = true;
-
-    // `description` is only valid as an explicit child of the components that
-    // declare a `description`/`descriptions` child group; it should not appear
-    // as a generic block child everywhere.
-    static inSchemaOnlyInheritAs = [];
-
-    static returnChildGroups() {
-        return [
-            {
-                group: "anything",
-                componentTypes: ["_base"],
-            },
-        ];
-    }
-
-    static returnStateVariableDefinitions() {
-        let stateVariableDefinitions = super.returnStateVariableDefinitions();
-
-        Object.assign(
-            stateVariableDefinitions,
-            returnPassThroughListItemChildStateVariableDefinitions(),
-        );
-
-        return stateVariableDefinitions;
-    }
-
-    recordVisibilityChange({ isVisible }) {
-        this.coreFunctions.requestRecordEvent({
-            verb: "visibilityChanged",
-            object: {
-                componentIdx: this.componentIdx,
-                componentType: this.componentType,
-            },
-            result: { isVisible },
-        });
-    }
 }
 
 export class XLabel extends Label {
